fix các chức năng cũ ở sprint 2

+++ src/main/java/com/be_mxh/repository/CommentRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 public interface CommentRepository
         extends JpaRepository<Comment, Long> {
 
-    long countByStatusId(Long statusId);
+    Integer countByStatusId(Long statusId);
 }
